+++ EDDI/EDDI.cs
@@ -1124,21 +1124,19 @@ namespace Eddi
 
         private bool eventEnteredNormalSpace(EnteredNormalSpaceEvent theEvent)
         {
+            // We won't update CurrentStation with this event, as doing so triggers false / premature updates from the Frontier API
+            CurrentStation = null;
+
             Environment = Constants.ENVIRONMENT_NORMAL_SPACE;
 
             if (theEvent.bodytype.ToLowerInvariant() == "station")
             {
                 // In this case body == station 
-                CurrentStellarBody = null;
             }
             else if (theEvent.body != null)
             {
-                // If we are not at a station then our station information is invalid 
-                CurrentStation = null;
-
                 // Update the body 
-                Logging.Debug("Now at body " + theEvent.body);
-                Body body = CurrentStarSystem.bodies.Find(s => s.name == theEvent.body);
+                Body body = CurrentStarSystem?.bodies?.Find(s => s.name == theEvent.body);
                 if (body == null)
                 {
                     // This body is unknown to us, might not be in EDDB or we might not have connectivity.  Use a placeholder 
@@ -1146,10 +1144,8 @@ namespace Eddi
                     body.name = theEvent.body;
                     body.systemname = theEvent.system;
                 }
-
                 CurrentStellarBody = body;
             }
-
             updateCurrentSystem(theEvent.system);
             return true;
         }
@@ -1319,11 +1315,13 @@ namespace Eddi
 
         private bool eventNearSurface(NearSurfaceEvent theEvent)
         {
+            // We won't update CurrentStation with this event, as doing so triggers false / premature updates from the Frontier API
+            CurrentStation = null;
+
             if (theEvent.approaching_surface)
             {
-                // Update the body we are approaching 
-                Logging.Debug("Now at body " + theEvent.body);
-                Body body = CurrentStarSystem.bodies.Find(s => s.name == theEvent.body);
+                // Update the body 
+                Body body = CurrentStarSystem?.bodies?.Find(s => s.name == theEvent.body);
                 if (body == null)
                 {
                     // This body is unknown to us, might not be in EDDB or we might not have connectivity.  Use a placeholder 
@@ -1331,7 +1329,6 @@ namespace Eddi
                     body.name = theEvent.body;
                     body.systemname = theEvent.system;
                 }
-
                 CurrentStellarBody = body;
             }
             else
@@ -1339,9 +1336,7 @@ namespace Eddi
                 // Clear the body we are leaving 
                 CurrentStellarBody = null;
             }
-
             updateCurrentSystem(theEvent.system);
-
             return true;
         }
 
@@ -1489,7 +1484,7 @@ namespace Eddi
                     // Save a timestamp when the API refreshes, so that we can compare whether events are more or less recent
                     ApiTimeStamp = DateTime.UtcNow;
 
-                    long profileTime = (long)DateTime.Now.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
+                    long profileTime = (long)DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
                     Profile profile = CompanionAppService.Instance.Profile();
                     if (profile != null)
                     {
@@ -1787,7 +1782,7 @@ namespace Eddi
 
                         // We do need to fetch an updated profile; do so
                         ApiTimeStamp = DateTime.UtcNow;
-                        long profileTime = (long)DateTime.Now.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
+                        long profileTime = (long)DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
                         Logging.Debug("Fetching station profile");
                         Profile profile = CompanionAppService.Instance.Station(CurrentStarSystem.name);
 
@@ -1811,7 +1806,7 @@ namespace Eddi
                             StarSystemSqLiteRepository.Instance.SaveStarSystem(CurrentStarSystem);
 
                             // Post an update event
-                            Event @event = new MarketInformationUpdatedEvent(DateTime.Now);
+                            Event @event = new MarketInformationUpdatedEvent(DateTime.UtcNow);
                             eventHandler(@event);
 
                             profileUpdateNeeded = false;
@@ -1846,7 +1841,7 @@ namespace Eddi
         private void dummyRefreshMarketData()
         {
             Thread.Sleep(2000);
-            Event @event = new MarketInformationUpdatedEvent(DateTime.Now);
+            Event @event = new MarketInformationUpdatedEvent(DateTime.UtcNow);
             eventHandler(@event);
         }