+++ SpeechResponder/ScriptResolver.cs
@@ -643,6 +643,32 @@ namespace EddiSpeechResponder
                 return (result == null ? new ReflectionValue(new object()) : new ReflectionValue(result));
             }, 1);
 
+            store["CommodityMarketDetails"] = new NativeFunction((values) =>
+            {
+                CommodityMarketQuote result = null;
+                if (values.Count == 1)
+                {
+                    // Named commodity, current station
+                    Station station = EDDI.Instance.CurrentStation;
+                    result = station == null ? null : station.commodities.FirstOrDefault(c => c.definition.localizedName == values[0].AsString);
+                }
+                else if (values.Count == 2)
+                {
+                    // Named commodity, named station, current system 
+                    StarSystem system = EDDI.Instance.CurrentStarSystem;
+                    Station station = system != null && system.stations != null ? system.stations.FirstOrDefault(v => v.name == values[1].AsString) : null;
+                    result = station == null ? null : station.commodities.FirstOrDefault(c => c.definition.localizedName == values[0].AsString);
+                }
+                else if (values.Count == 3)
+                {
+                    // Named commodity, named station, named system 
+                    StarSystem system = DataProviderService.GetSystemData(values[2].AsString, null, null, null);
+                    Station station = system != null && system.stations != null ? system.stations.FirstOrDefault(v => v.name == values[1].AsString) : null;
+                    result = station == null ? null : station.commodities.FirstOrDefault(c => c.definition.localizedName == values[0].AsString);
+                }
+                return (result == null ? new ReflectionValue(new object()) : new ReflectionValue(result));
+            }, 0, 3);
+
             store["BlueprintDetails"] = new NativeFunction((values) =>
             {
                 BlueprintMaterials result = BlueprintMaterials.FromName(values[0].AsString);